Rebasing master and using the same type on width and height variables and parameters

+++ Asteroids/Game.cpp
@@ -10,7 +10,7 @@
 
 namespace Asteroids
 {
-	Game::Game(const std::string& title, const float width, const float height)
+	Game::Game(const std::string& title, const int width, const int height)
 		: m_title(title),
 		m_width(width),
 		m_height(height)
